Salon Management  System
Salon management system is a one of the desktop base application. This 
application used by all small and medium shop. It is used to keep and manage 
the customer details along with purchase and bill details. Salon management 
system is easy to use and completely installation based. Application also 
provide the add-on facility of manage supplier details.